refactor(sdk-lib-mpc): move ed25519 and shamir test
TICKET: WP-400
