+++ modules/sdk-lib-mpc/src/curves/index.ts
@@ -1,5 +1,5 @@
 export * from './types';
+export * from './util';
+export * from './bip32HdTree';
 export * from './secp256k1';
-export * from './bip32';
-export * from './hdTree';
-export * from './shamir';
+export * from './ed25519';
