+++ modules/sdk-lib-mpc/src/index.ts
@@ -2,3 +2,4 @@ export * from './tss';
 export * from './curves';
 export * from './util';
 export * from './openssl';
+export * from './shamir';
