+++ modules/sdk-lib-mpc/src/shamir/types.ts
@@ -1,0 +1,4 @@
+export type SplitSecret = {
+  shares: Record<number, bigint>;
+  v: bigint[];
+};
